feat: update landing booking and searchbar and formatdate utils
