+++ client/src/pages/SearchResultPage.jsx
@@ -1,13 +1,13 @@
 import SearchResult from "../components/SearchResult";
 import NavigationBar from "../components/navigation-bar/NavigationBar";
 import Footer from "../components/Footer";
-import LandingBooking from "../components/landing-page/LandingBooking";
+import SearchBar from "../components/utils/SearchBar";
 
 function SearchResultPage() {
   return (
     <>
       <NavigationBar />
-      <LandingBooking />
+      <SearchBar />
       <SearchResult />
       <Footer />
     </>
